log: rearrange .go files
