+++ pkg/log/log.go
@@ -33,28 +33,6 @@ func New(cloudProvider, level string) *Logger {
 	return &Logger{Logger: l}
 }
 
-// Entry is an abstraction to the
-// Logger and the logrus.Entry
-// so that *Logger always creates
-// an Entry copy which ensures no
-// Fields are being overwritten.
-type Entry interface {
-	// Basic Logging Operation
-	Debugf(format string, args ...interface{})
-	Infof(format string, args ...interface{})
-	Warnf(format string, args ...interface{})
-	Errorf(format string, args ...interface{})
-
-	// Attach contextual information to the logging entry
-	WithFields(fields map[string]interface{}) Entry
-
-	// SystemErr is a method that disects the error
-	// and logs the appropriate level and fields for it.
-	// TODO(marwan-at-work): When we have our own Error struct
-	// this method will be very useful.
-	SystemErr(err error)
-}
-
 // SystemErr Entry implementation.
 func (l *Logger) SystemErr(err error) {
 	l.Logger.Error(err)
